Investigation of N=100 analog bias experiment.
parenting18analogsAsBiases1b.lisp has popsize N=100

Changed the big NetLogo programs to remove old references
to the goo and profiler libraries, which are not needed at present.

The new readme file points to instructions to change the Java heap
size, necessary to run the "Humongous" NetLogo program with a huge
world window.