Last night's revision to work around meta cols. However, by the end of this work on df2ra.R, I became convinced by other exploration that it will be simpler for now to leave out the meta columns runid and tick.  That's why there are also commits today removing these columns from the Lisp code, and simplifying output file names to consist mainly of the runid plus extension.
Tick # is implicit in my data formats in R, and was normally redundant.
Taking it out does mean that you have to be careful if you subset a
dataset so that it starts at a tick other than 1.

As for runid, although I do want to keep track of which run each dataset
is from in R, having an entire column in which the runid is repeated
for every time step (tick) is turning out to be more trouble than help.
Instead, I'm going to normally name the csv files in a simple, standard way
using the runid.  Then I can have code in R that extracts the runid
for a dataset from the filename from which the dataset originates.
A little kludgey--depending on filenames--but any way of including the runid
in the data file itself complicates reading in and processing the data,
and it's just not worth all that, for present purposes at least.